+++ src/types/models/Keyword.ts
@@ -1,5 +1,0 @@
-export enum Keyword {
-  CreatedAt = '$[createdAt]',
-  UpdatedAt = '$[createdAt]',
-  DeletedAt = '$[createdAt]',
-}