+++ src/types/models/delete.ts
@@ -1,0 +1,8 @@
+export interface DeleteSetting {
+  isDefaultRecursive: boolean
+  isDefaultOpenCard: boolean
+  isNotFirstPage: boolean
+}
+
+export type DeleteSettingUpdate = Partial<DeleteSetting>
+