+++ src/types/models/move.ts
@@ -7,6 +7,8 @@ export interface MoveSetting {
   isAutoDuplicatedName: boolean
   isKeepOriginal: boolean
   isDefaultCheckedOnLoad: boolean
+  isNotFirstPage: boolean
+  isNotFirstLoad: boolean
 }
 
 export type MoveSettingUpdate = Partial<MoveSetting>